// SETPPBO/ClientApp/app/main/administrasi-banding-gugatan/models/penyampaian-salinan-bantahan.ts
export interface PenyampaianSalinanBantahan {
    PermohonanId: string;
    NoSuratPermintaanSalinan: string;
    TglSuratPermintaanSalinan: Date;
    NoSengketa: string;
    RefTermohonId: number;
    RefUpTermohonId: number;
    NamaTermohon: string;
    NamaUpTermohon: string;
    AlamatTermohon: string;
    RefKotaTermohonId: number;
    KotaTermohon: string;
    NoSuratBantahan: string;
    TglSuratBantahan: Date;
    NoSuratPermohonan: string;
    TglSuratPermohonan: Date;
    RefTtdPermintaanSalinanId: number;
    RefStatusId: number;
}