// SETPPBO/ClientApp/app/main/referensi/models/ref-kode-termohon.ts
export interface RefKodeTermohon {
    OrganisasiId: number;
    IndukOrganisasiId: number;
    KodeOrganisasi: string;
    Level: number;
    UraianOrganisasi: string;
    UraianLengkapOrganisasi: string;
    UraianJabatan: string;
    KodeSatker: string;
    Alamat: string;
    Kota: string;
    KodePos: string;
    KodeSurat: string;
}